--- src/components/DatePicker.tsx
@@ -1,7 +1,0 @@
-interface Props {}
-
-const DatePicker = (props: Props) => {
-  return <div>DatePicker</div>
-}
-
-export default DatePicker
